Refactor center position
- Additional refactor to allow structuring utils
