+++ src/napari_spatialdata/_scatterwidgets.py
@@ -20,8 +20,8 @@ from qtpy.QtCore import Signal
 
 from napari_spatialdata._categoricals_utils import _add_categorical_legend
 from napari_spatialdata._model import ImageModel
-from napari_spatialdata._utils import NDArrayA, _get_categorical, _set_palette
 from napari_spatialdata._widgets import AListWidget, ComponentWidget
+from napari_spatialdata.utils._utils import NDArrayA, _get_categorical, _set_palette
 
 __all__ = [
     "MatplotlibWidget",
