+++ src/napari_spatialdata/utils/_utils.py
